% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Activity}
\usage{
activity(activity_data, gps_data = NULL, parameters = NULL,
  keep_source = TRUE)
}
\value{
An object of class \code{activity}.
}
\description{
\code{activity()} creates an object of class \code{activity}
}
\examples{
  activity()
}
