% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2thresholds.R
\name{activity2thresholds}
\alias{activity2thresholds}
\title{Activity2thresholds}
\usage{
activity2thresholds(activity, activity_gaps, act, act_ma, reg_minutes,
  n_runs, window_width_around_day, n_thresholds, resting_range_limit,
  threshold_range_limit, min_duration_active_state, plot_summary)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag' and 'ts'.}

\item{activity_gaps}{A data.table with the identified gaps in the activity
data (return of function \code{\link{identify_activity_gaps}}.}

\item{act}{a}

\item{act_ma}{a}

\item{n_runs}{a}

\item{window_width_around_day}{a}

\item{n_thresholds}{a}

\item{resting_range_limit}{a}

\item{threshold_range_limit}{a}

\item{min_duration_active_state}{a}

\item{plot_summary}{}
}
\value{
Data.table with the ...
}
\description{
\code{activity2thresholds}
}
\examples{
activity_thresholds <- activity2thresholds(activity = activity_data,
                              activity_gaps = activity_data_gaps,
                              act = 'act_xy',
                              act_ma = 'act_xy_ma2',
                              reg_minutes = 5,
                              n_runs = 1,
                              window_width_around_day = 3,
                              n_thresholds = c(25:35),
                              resting_range_limit = 0.2,
                              threshold_range_limit = 0.8,
                              min_duration_active_state = 2,
                              plot_summary = TRUE)

}
