% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_data}
\alias{activity_data}
\title{Activity data.}
\format{A data frame with 5 variables and 301262 rows:
\describe{
  \item{animal_id}{identifier of the animal.}
  \item{tag_code}{identifier of the tag.}
  \item{act_x}{activity values of the x-axis.}
  \item{act_y}{activity values of the y-axis.}
  \item{ts}{Timestamp of the activity measurement at time zone UTC.}
}}
\usage{
activity_data
}
\description{
A data set containing the activity data of three tagged red deer. The sensor
 measured acceleration in two axis (x = forward–backward horizontal motion
 and y =  left–right horizontal motion) and was set to record activity
 every 5 minutes.
}
\keyword{datasets}
