% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gps_data.R
\name{add_gps_data}
\alias{add_gps_data}
\title{Add GPS data}
\usage{
add_gps_data(x, gps_data)
}
\value{
An object of class \code{activity} or \code{states}
}
\description{
\code{add_gps_data()} adds or changes the \code{$gps_data} of a \code{states} or
  \code{activity} object.
}
\examples{
  states()
}
