% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_thresholds.R
\name{aggregate_thresholds}
\alias{aggregate_thresholds}
\title{Aggregate activity thresholds}
\usage{
aggregate_thresholds(thresholds)
}
\arguments{
\item{thresholds}{A data.table with the activity thresholds (return of
\code{\link{activity2thresholds}} with additional column
\code{threshold_period)} by which to aggregate the threshold values.}
}
\description{
\code{aggregate_thresholds} aggregates the daily activity thresholds (return
  of \code{\link{activity2thresholds}} by \code{period}.
}
\examples{
aggregate_tresholds(thresholds = activity_thresholds_final)

}
