% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_states.R
\name{calculate_states}
\alias{calculate_states}
\title{Calculate activity states}
\usage{
calculate_states(activity, thresholds = c("a", "b", "c"), add = NULL,
  pta.pos = NULL, pta.dayshift = NULL, pta.dawn_degree = NULL,
  pta.period = NULL, pta.max_na = NULL)
}
\value{
An object of class \code{activity}.
}
\description{
\code{calculate_states()} calculates active states for an \code{activity} object.
}
\examples{
  calculate_states()
}
