% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_thresholds.R
\name{calculate_thresholds}
\alias{calculate_thresholds}
\title{Calculate activity thresholds}
\usage{
calculate_thresholds(activity, act.act = NULL, act.reg_minutes = NULL,
  thresh.n_runs = 1, thresh.window_width_around_day = NULL,
  thresh.n_thresholds = NULL, thresh.resting_range_limit = 0.2,
  thresh.threshold_range_limit = 0.8,
  states.min_duration_active = NULL, plot_summary = FALSE)
}
\arguments{
\item{activity}{An object of class \code{activity}.}
}
\value{
An object of class \code{activity}.
}
\description{
\code{calculate_thresholds} calculates and aggregates activity thresholds
  for an \code{$activity} object, based on its \code{$activity_data} and
  \code{$activity_gaps}. Raw thresholds (per day) are stored in
  \code{$activity_data} and aggregate thresholds are stored in
  \code{$activity_data}
}
\examples{
  calculate_thresholds()
}
