% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_animaltag.R
\name{create_animaltag}
\alias{create_animaltag}
\title{Add column animal_tag}
\usage{
create_animaltag(data)
}
\arguments{
\item{data}{A data.table with the columns 'animal_id' and 'tag_code'.}
}
\value{
The original data.table with additional column 'animal_tag'.
}
\description{
\code{create_animaltag}  concatenates the columns 'animal_id' and
  'animal_tag' as identifier for the animal-tag-association.
}
\examples{
activity_data <- create_animaltag(activity_data)
}
