% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters}
\usage{
get_parameters(x, parameters)
}
\value{
A named list with containing the values of the specified parameters
}
\description{
\code{get_poarameters} extracts specified parameters from objects of class
  \code{activity} or \code{states}. An error is caused if any of the
  parameters is \code{NA}.
}
