% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gps_data}
\alias{gps_data}
\title{GPS data.}
\format{A data frame with 6 variables and 15780 rows:
\describe{
  \item{animal_id}{Identifier of the animal.}
  \item{tag_code}{Identifier of the tag.}
  \item{longitude}{Longitude coordinate of the GPS position in WGS84
        (EPSG:4326).}
  \item{latitude}{Latitude coordinate of the GPS position in WGS8
        (EPSG: 4326).}
  \item{dop}{Dilution of precision of the GPS position.}
  \item{ts}{Timestamp of the GPS position.}
}}
\usage{
gps_data
}
\description{
A data set containing the GPS data of three tagged red deer. The sensor was
set to acquire a GPS position every two hours.
}
\keyword{datasets}
