% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_activity_gaps.R
\name{identify_activity_gaps}
\alias{identify_activity_gaps}
\title{Identify gaps in the activity data}
\usage{
identify_activity_gaps(activity, act.act = NULL)
}
\arguments{
\item{activity}{An object of class \code{activity}.}

\item{act}{The activity variable to scan for gaps.}
}
\value{
An object of class \code{activity} containing \code{$activity_gaps}
  in which the start- and end-timestamps of the identified gaps are
  specified.
}
\description{
\code{identify_activity_gaps} searches for data gaps in the
  \code{$activity_data} of an \code{activity} object, then stores the gaps
  in the object as \code{$activity_gaps}
}
\examples{
activity_data_gaps <- identify_activity_gaps(activity = activity_data,
                                             act = 'act_xy')
}
