% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{plot_activity}
\alias{plot_activity}
\title{Plot activity data}
\usage{
plot_activity(activity, gps = NULL, animal_id = NULL, tag_code = NULL,
  axis, plot_gps, plot_movement, plot_moving_window, plot_state,
  window_width_activity_ma, threshold, min_duration_active_state,
  plot_predicted_state)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag' and 'ts'.}

\item{gps}{A data.table with the gps data.}

\item{threshold_parm}{}
}
\description{
\code{plot_activity} plots activity data against time of day for each day of
  the week.
}
\examples{
plot_activity(activity = activity_data,
              animal_id = '12',
              gps = gps_data)

}
