% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity_eval.R
\name{plot_activity_eval}
\alias{plot_activity_eval}
\title{Plot activity data}
\usage{
plot_activity_eval(activity, gps, act.available_act,
  select_animal_id = NULL)
}
\arguments{
\item{gps}{A data.table with the gps data.}

\item{select_animal_id}{}
}
\description{
\code{plot_activity} plots activity data against time of day for each day of
  the week.
}
\examples{

}
