% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_states.R
\name{plot_states}
\alias{plot_states}
\title{Plot states}
\usage{
plot_states(active_states, prop_time_active, gps = NULL)
}
\arguments{
\item{states}{An object of class \code{activity}.}

\item{threshold}{Type of thresholds. Must be one of \code{"a"}, \code{"b"}, \code{"c"}.}
}
\description{
\code{plot_states} plots the active states and proportion of time in state
  active for each animal.
}
\examples{
plot_states(active_states = active_states_a,
            prop_time_active = prop_time_active,
            gps = gps_data)

}
