% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thresholds.R
\name{plot_thresholds}
\alias{plot_thresholds}
\title{Plot thresholds}
\usage{
plot_thresholds(thresholds)
}
\arguments{
\item{thresholds}{}
}
\description{
\code{plot_thresholds} plots the calculated activity thresholds against the
 period by which thersholds are aggregated.
}
\examples{
plot_tresholds(thresholds = activity_thresholds_final,
               period = 'year_week')
}
