% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_activity.R
\name{regularize_activity}
\alias{regularize_activity}
\title{Regularize activity data}
\usage{
regularize_activity(activity, minutes)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag', 'ts', 'act_x' and 'act_y'.}

\item{minutes}{The regular interval in minutes.}
}
\value{
The original data.table with regularized timestamps in column 'ts'.
}
\description{
\code{regularize activity} regularizes the timestamps in column 'ts' to
  regular intervals for each animal_tag. The interval is defined by the
  parameter \code{minutes}. The following steps are performd:
  \enumerate{
    \item Round each timestamp to the nearsed regular timestamp.
    \item When several data points occur with the same regularized timestamp,
       aggregate these by calculation the mean of the activity values.
    \item Insert data points for missing regularized timestamps with
      NA-activity values .
    }
}
\examples{
activity_data <- regularize_activity(data = activity_data, minutes = 5)
}
