% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters for object of class \code{activity}}
\usage{
set_parameters(x, parameters)
}
\value{
An object of class \code{activity} or \code{states}
}
\description{
\code{set_poarameters} takes a named list of parameter values and sets the
matching parameters in an object of class \code{activity} or \code{states} to the
corresponding values. (i.e. \code{NA})
}
