% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_animaltag.R
\name{split_animaltag}
\alias{split_animaltag}
\title{Split column animal_tag}
\usage{
split_animaltag(data)
}
\arguments{
\item{data}{A data.table with the column 'animal_tag'.}
}
\value{
The original data.table with additional columns 'animal_id' and
  'tag_code'.
}
\description{
\code{split_animaltag}  splits the column animal_tag in the columns
  'animal_id' and 'animal_tag' and adds them to the table.
}
\examples{
activity_data <- split_animaltag(activity_data)
}
