% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2prop_time_active.R
\name{states2prop_time_active}
\alias{states2prop_time_active}
\title{Calculate proportion of time in state active}
\usage{
states2prop_time_active(active_states, activity_gaps, gps = NULL,
  pos = NULL, dayshift = "sunrise", dawn_degree = 12, period = "day",
  max_na = 0)
}
\arguments{
\item{active_states}{A data.table with the predicted active states (return
of function \code{\link{thresholds2states}}.}

\item{activity_gaps}{A data.table with the identified gaps in the activity
data (return of function \code{\link{identify_activity_gaps}}.}

\item{gps}{A data.table with the GPS positions (at least one) of the animals.
The table should contain the columns: <animal_tag>, <longitude>, <latitude>.}

\item{pos}{A vector containing longitude and latitude of the research area.
This parameter is only necessary when no GPS-positions are provided or when
the GPS-positions for some animals are missing.}

\item{dayshift}{Either "sunrise" or "dawn". Defines when consecutive days are
separated so that each day contains one complete night and one complete day.
The identifier for each day is called <date_se> for sunrise and <date_dawn>
for dawn.}

\item{dawn_degree}{An integer between 0 and 90 that defines the angle of the
sun below the horizon at dawn and dusk.}

\item{period}{Either "day" or "week", specifies the period of time in which the
proportion of time in state active is aggregated.}

\item{max_na}{An integer with the maximum number of minutes with NA
that are alloes for each period. When this threshold is crossed proportion of
time in state active will be set to NA for this period.}
}
\value{
A data.table with the proportion of time in state active for each
  animal_tag and day. The proportion of time in state active is returned for
  the whole day <total>, during daytime <day> (sunrise till sunset), during
  nighttime <night> (dusk till dawn), during dusk <dusk> (sunset till the
  sun reaches dawn_degree),during dawn <dawn> ()
}
\description{
\code{states2prop_time_active} calculates the proportion of time in state
  active for each animal_tag and day.
}
\examples{
states2prop_time_active(active_states = active_states_a,
                                  activity_gaps = activity_data_gaps,
                                  gps = gps_data,
                                  dayshift = "sunrise",
                                  dawn_degree = 12,
                                  period = "day"
                                  max_na = 30)

}
