% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds2states.R
\name{thresholds2states}
\alias{thresholds2states}
\title{Thresholds2states}
\usage{
thresholds2states(parameters, activity, activity_gaps, thresholds,
  threshold_par)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag' and 'ts'.}

\item{activity_gaps}{The data.table with the identified activity gaps.}

\item{threshold_parm}{}
}
\value{
Data.table with the ...
}
\description{
Calculates active states based on a list of threshold values (return
  of function \code{\link{activity2thresholds}} that is grouped by a
  defined period with function \code{\link{aggregate_thresholds}})
}
\examples{
active_states_a <-  thresholds2states(activity = activity_data,
                                      activity_gaps = activity_data_gaps,
                                      thresholds = activity_thresholds_final,
                                      threshold_par = 'threshold_a')

}
