% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2minutes.R
\name{ts2minutes}
\alias{ts2minutes}
\title{Timestamp 2 minutes}
\usage{
ts2minutes(x)
}
\arguments{
\item{ts}{A vector of timestamps with class POSIXct.}
}
\value{
A integer vector with the elapsed minutes since 00:00.
}
\description{
\code{ts2minutes} calculates the elapsed minutes of the day from 00:00.
}
\examples{


}
