% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2states.R
\name{activity2states}
\alias{activity2states}
\title{Activity2states}
\usage{
activity2states(activity, activity_gaps = NULL, axis = "act_xy",
  axis_ma = "no", width_axis_ma, threshold, min_duration_active_state)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag' and 'ts'.}

\item{activity_gaps}{The data.table with the identified activity gaps.}

\item{min_duration_active_state}{}
}
\description{
\code{activity2states}
}
\examples{


}
