% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2thresholds.R
\name{activity2thresholds}
\alias{activity2thresholds}
\title{Activity2thresholds}
\usage{
activity2thresholds(activity, activity_gaps, axis, axis_ma, n_runs,
  window_width_around_day, n_thresholds, min_bin_width,
  min_duration_active_state, plot_summary)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag' and 'ts'.}

\item{activity_gaps}{A data.table with the identified gaps in the activity
data (return of function \code{\link{identify_activity_gaps}}.}

\item{axis}{a}

\item{axis_ma}{a}

\item{n_runs}{a}

\item{window_width_around_day}{a}

\item{n_thresholds}{a}

\item{min_bin_width}{a}

\item{min_duration_active_state}{a}

\item{plot}{summary}
}
\value{
Data.table with the ...
}
\description{
\code{activity2thresholds}
}
\examples{
activity_thresholds <- activity2thresholds(activity = activity_data,
                              activity_gaps = activity_data_gaps,
                              axis = 'act_xy',
                              axis_ma = 'act_xy_ma2',
                              n_runs = 1,
                              window_width_around_day = 3,
                              n_thresholds = c(25:35),
                              min_bin_width = 1,
                              min_duration_active_state = 10,
                              plot_summary = TRUE)

}
