% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_activity_gaps.R
\name{identify_activity_gaps}
\alias{identify_activity_gaps}
\title{Identify data gaps in the activity data}
\usage{
identify_activity_gaps(activity, axis)
}
\arguments{
\item{activity}{A data.table with the activity data. The following columns
should be present: 'animal_tag', 'ts' and the activity variable that is
scanned for gaps (e.g. 'act_x').}

\item{axis}{The activity variable to scan for gaps.}
}
\value{
A data.table with the start- and end-timestamps of the identified
  gaps.
}
\description{
\code{identify_activity_gaps} searches for data gaps in the activity data.
}
\examples{
activity_data_gaps <- identify_activity_gaps(activity = activity_data,
                                             axis = 'act_xy')
}
