% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_states.R
\name{plot_states}
\alias{plot_states}
\title{Plot states}
\usage{
plot_states(active_states, prop_time_active, gps = NULL)
}
\arguments{
\item{active_states}{A data.table with the active states (return of function
\code{\link{thresholds2states}}).}

\item{prop_time_active}{A data.table with the proportion of time in state
active (return of \code{\link{states2prop_time_active}})}

\item{gps}{A data.table with the GPS-positions of the animals.}
}
\description{
\code{plot_states} plots the active states and proportion of time in state
  active for each animal.
}
\examples{
plot_states(active_states = active_states_a,
            prop_time_active = prop_time_active,
            gps = gps_data)

}
