% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_activity.R
\name{regularize_activity}
\alias{regularize_activity}
\title{Regularize activity data}
\usage{
regularize_activity(activity, act.reg_minutes = NULL)
}
\arguments{
\item{activity}{An object of class \code{activity}.}

\item{act.reg_minutes}{The regular interval in minutes.}
}
\value{
The original activity object with regularized timestamps for its
  \code{$acticity_data} in column 'ts'.
}
\description{
\code{regularize_activity} regularizes the timestamps within the
  \code{$acticity_data} of an \code{activity} object (column 'ts') to
  regular intervals for each animal_tag. The interval is defined by the
  parameter \code{minutes}. The following steps are performed:
}
\details{
\enumerate{
    \item Round each timestamp to the nearsed regular timestamp.
    \item When several data points occur with the same regularized timestamp,
       aggregate these by calculating the mean of the activity values.
    \item Insert data points for missing regularized timestamps with
      NA-activity values .
    }
}
\examples{
activity_data <- regularize_activity(data = activity_data, minutes = 5)
}
