% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_activity_gaps.R
\name{remove_activity_gaps}
\alias{remove_activity_gaps}
\title{Remove activity gaps}
\usage{
remove_activity_gaps(activity, activity_gaps)
}
\arguments{
\item{activity}{An object of class \code{activity}.}
}
\description{
\code{remove_activity_gaps} removes data gaps in the \code{$activity_data}
  of an \code{activity} object. The data gaps have to be identified before
  with the function \code{\link{identify_activity_gaps}}.
}
\examples{
activity_data <- remove_activity_gaps(activity = activity_data,
                                      activity_gaps = activity_data_gaps)
}
