% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_activity.R
\name{smooth_activity}
\alias{smooth_activity}
\title{Smooth activity data}
\usage{
smooth_activity(activity, act, act.width_ma, update_NA = TRUE)
}
\value{
A data.table
}
\description{
\code{smooth_activity_data} smoothes an activity variable by calculating a
  moving average with window width specified by \code{act.width_ma}.
}
\examples{
activity_data <- smooth_activity_data(activity = activity_data,
                                          act = 'act_xy',
                                          act.width_ma = 2,
                                          update_NA = FALSE)
}
